backend/requirements.txt
firebase-admin==7.1.0
mysql-connector-python==9.5.0
pymongo==4.6.1
motor==3.3.2  # Driver MongoDB asynchrone pour FastAPI
dnspython==2.4.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# LangChain & Groq (compatible OpenAI) - Versions compatibles testées
# Groq utilise l'interface OpenAI compatible, donc on utilise langchain-openai avec base_url
langchain==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-community==0.2.16
openai>=1.40.0,<2.0.0  # Version compatible avec langchain-openai 0.1.23 (utilisé pour Groq via base_url)
langsmith==0.1.129

# RAG & Vector Stores
chromadb==0.4.22  # Alternative vector store optionnel
tiktoken>=0.7,<1  # Compatible avec langchain-openai 0.1.23 (doit être >=0.7)
sentence-transformers>=2.2.0  # Pour les embeddings multilingues
numpy>=1.24.0  # Pour les calculs vectoriels
tqdm>=4.65.0  # Pour les barres de progression

